+++ packages/framer-motion-3d/src/render/__tests__/index.test.tsx
@@ -58,6 +58,7 @@ describe("motion for three", () => {
             ReactThreeTestRenderer.create(<Component />)
         })
 
+        expect(result.length).toBeGreaterThan(1)
         const lastFrame = result[result.length - 1]
         expect(lastFrame).toEqual({
             scale: 5,
@@ -230,8 +231,8 @@ describe("motion for three", () => {
                                 transition: { x: { type: false } },
                             },
                             off: {
-                                x: 0,
-                                y: 0,
+                                x: 1,
+                                y: 1,
                                 transition: { x: { type: false } },
                             },
                         }}
@@ -240,10 +241,10 @@ describe("motion for three", () => {
             }
 
             ReactThreeTestRenderer.create(<Component />)
-            frame.update(() => resolve([x.get(), y.get()]))
+            frame.postRender(() => resolve([x.get(), y.get()]))
         })
 
         expect(result[0]).toEqual(100)
-        expect(result[1]).toEqual(0)
+        expect(result[1]).not.toEqual(100)
     })
 })