+++ packages/framer-motion/src/animation/animators/waapi/utils/get-final-keyframe.ts
@@ -1,12 +1,16 @@
 import { Repeat } from "../../../../types"
 
+const isNotNull = (value: unknown) => value !== null
+
 export function getFinalKeyframe<T>(
     keyframes: T[],
     { repeat, repeatType = "loop" }: Repeat
 ): T {
+    const resolvedKeyframes = keyframes.filter(isNotNull)
     const index =
         repeat && repeatType !== "loop" && repeat % 2 === 1
             ? 0
-            : keyframes.length - 1
-    return keyframes[index]
+            : resolvedKeyframes.length - 1
+
+    return resolvedKeyframes[index]
 }