+++ packages/framer-motion/src/gestures/__tests__/focus.test.tsx
@@ -1,10 +1,11 @@
 import { focus, blur, render } from "../../../jest.setup"
 import * as React from "react"
 import { motion, motionValue } from "../../"
+import { nextFrame } from "./utils"
 
 describe("focus", () => {
     test("whileFocus applied", async () => {
-        const promise = new Promise((resolve) => {
+        const promise = new Promise(async (resolve) => {
             const opacity = motionValue(1)
             const ref = React.createRef<HTMLAnchorElement>()
             const Component = () => (
@@ -24,6 +25,8 @@ describe("focus", () => {
             ref.current!.matches = () => true
 
             focus(container, "myAnchorElement")
+
+            await nextFrame()
 
             resolve(opacity.get())
         })
@@ -60,7 +63,7 @@ describe("focus", () => {
     })
 
     test("whileFocus applied if focus-visible selector throws unsupported", async () => {
-        const promise = new Promise((resolve) => {
+        const promise = new Promise(async (resolve) => {
             const opacity = motionValue(1)
             const ref = React.createRef<HTMLAnchorElement>()
             const Component = () => (
@@ -87,6 +90,8 @@ describe("focus", () => {
 
             focus(container, "myAnchorElement")
 
+            await nextFrame()
+
             resolve(opacity.get())
         })
 
@@ -95,7 +100,7 @@ describe("focus", () => {
 
     test("whileFocus applied as variant", async () => {
         const target = 0.5
-        const promise = new Promise((resolve) => {
+        const promise = new Promise(async (resolve) => {
             const variant = {
                 hidden: { opacity: target },
             }
@@ -120,6 +125,8 @@ describe("focus", () => {
 
             focus(container, "myAnchorElement")
 
+            await nextFrame()
+
             resolve(opacity.get())
         })
 
@@ -127,7 +134,7 @@ describe("focus", () => {
     })
 
     test("whileFocus is unapplied when blur", () => {
-        const promise = new Promise((resolve) => {
+        const promise = new Promise(async (resolve) => {
             const ref = React.createRef<HTMLAnchorElement>()
             const variant = {
                 hidden: { opacity: 0.5, transitionEnd: { opacity: 0.75 } },
@@ -157,6 +164,7 @@ describe("focus", () => {
             ref.current!.matches = () => true
 
             focus(container, "myAnchorElement")
+            await nextFrame()
             setTimeout(() => {
                 blurred = true
                 blur(container, "myAnchorElement")